Remove support for setting sandbox flags via "allow"
This removes the experimental FeaturePolicyForSandbox feature, which
allowed a document to set sandbox flags via Permissions Policy (then
Feature Policy). This feature is not going to ship, and the mechanism
being discussed for it in the future is Document Policy, rather than
Permissions Policy.

This also removes the "opener feature state" for popup windows, which
was only used by these features. The Permissions Policy spec does not
support this any more, as all feature state is reset in new top-level
browsing contexts.

Change-Id: Ifc58ce6095892f8c4b968cfc0c02bbe8060d985c