[LayoutNG] Abspos elements inside a spanner
It was decided that the containing block chain goes directly from a
column spanner to the multi-column container. Thus, if an absolute
positioned element is nested inside a spanner, we should begin
looking for its containing block starting with the multi-column
container, as well.

This is an update to the previous legacy behavior, and the related
web test has been updated to reflect the new behavior.

Bug: 1079031
Change-Id: Ia9f60107417158c32d78e12a1d414181411ecb0a